Move test to MMap dir, not all directories throw exception on seek
